fix: 인기있는 여행지 섹션 잘리는 문제 해결
